Various fixes
ToggleProductNeedsRefill from SqlConnector now returns the updated value from the database instead of changing it both places separately.

Updated the ListBox in EditProductsFrom to work similarly to in StartpageFrom.

Added new error message option for when tags are added.

Adding a new product will now automatically open the form to edit it.

Started new form to make it easy to edit the connections a specific tag has to all the products.